+++ .github/workflows/dev-containers.yml
@@ -41,9 +41,12 @@ jobs:
           "src/test/container-features/e2e.test.ts",
           "src/test/container-features/featuresCLICommands.test.ts",
           "src/test/container-features/containerFeaturesOrder.test.ts",
+          "src/test/cli.build.test.ts",
+          "src/test/cli.exec.test.ts",
           "src/test/cli.test.ts",
+          "src/test/cli.up.test.ts",
           # Run all except the above:
-          "--exclude src/test/container-features/e2e.test.ts --exclude src/test/container-features/featuresCLICommands.test.ts --exclude src/test/container-features/containerFeaturesOrder.test.ts --exclude src/test/cli.test.ts 'src/test/**/*.test.ts'",
+          "--exclude src/test/container-features/e2e.test.ts --exclude src/test/container-features/featuresCLICommands.test.ts --exclude src/test/container-features/containerFeaturesOrder.test.ts --exclude src/test/cli.build.test.ts --exclude src/test/cli.exec.test.ts --exclude src/test/cli.test.ts --exclude src/test/cli.up.test.ts 'src/test/**/*.test.ts'",
         ]
     steps:
     - name: Checkout
